placed validation schema into its own file and is being imported, making put request to api and updating context to reflect changes
